requirements.txt
pyyaml
pandas
pyfiglet
Pyarrow
tmdbv3api
redis
rq
ffmpeg-python